gh-660: Upgrade go-graphql and use thunk again for local get class
The only difference is that this time we're creating the thunk as part
of the resolver - not on the connector.
